Switch to the Beeper-maintained fork of mx-puppet-discord
Building mx-puppet-discord appears to have been broken for a while.

Related to:
- https://github.com/matrix-discord/mx-puppet-discord/issues/201
- https://github.com/matrix-discord/mx-puppet-discord/issues/202
- https://github.com/matrix-discord/mx-puppet-discord/issues/203

We'd rather use a fork that is maintained better and by someone who
cares about whether their software works or not, so we'll be using the
Beeper-maintained for from now on.

In the future, we should probably do the same for the Slack bridge
which is also part of the same monorepo
(https://gitlab.com/beeper/mx-puppet-monorepo).