Temporarily switch to Myned-maintained fork of mx-puppet-discord/mx-puppet-slack
Fixes https://github.com/spantaleev/matrix-docker-ansible-deploy/issues/1719

We're using a pinned commit of this fork, which patches up the npm/yarn
lock files, so that self-building of the Discord and Slack bridges can
work after GitHub's deprecation of the `git://` and `ssh://` protocols
(https://github.blog/2021-09-01-improving-git-protocol-security-github/).

When the issue gets fixed in the Beeper fork (via
https://gitlab.com/beeper/mx-puppet-monorepo/-/merge_requests/35 or
otherwise), we'll get back on it.