Adjust "installed using this playbook" links when linking to specific components
We previously made "this playbook" a link to some specific component.
This is somewhat misleading. Reading the text and seeing that "this playbook"
is a link, one might think that "this playbook" actually links to the
playbook's home page or something, not to the specific component which
could be "installed using this playbook".

This patch adjusts the markup so that the whole phrase "installed using this playbook"
is turned into a link (to the specific component's page). This makes it
more obvious where we're linking.

This work was initially done for the Email2Matrix docs page in
1b05343aba19852d627ded4934c0580960aaadb9. This patch brings the same
change to all other instances.